requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-Bootstrap==3.3.7.1
bootstrap-flask==2.3.3
pandas
openpyxl
xlsxwriter